+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix issue - Joined geometry does not convert correctly
+
 ## 0.1.12
 * Remove transient elements in Revit doc when Dynamo shutdown
 * Fix test failure of JsonRestoresBinding